+++ python/cudf/cudf/dataframe/columnops.py
@@ -209,7 +209,7 @@ def column_select_by_position(column, positions):
     )
 
 
-def build_column(buffer, dtype, mask=None, categories=None):
+def build_column(buffer, dtype, mask=None, categories=None, name=None):
     from cudf.dataframe import numerical, categorical, datetime, string
 
     if pd.api.types.is_categorical_dtype(dtype):
@@ -219,20 +219,23 @@ def build_column(buffer, dtype, mask=None, categories=None):
             categories=categories,
             ordered=False,
             mask=mask,
+            name=name,
         )
     elif np.dtype(dtype).type == np.datetime64:
         return datetime.DatetimeColumn(
-            data=buffer, dtype=np.dtype(dtype), mask=mask
+            data=buffer, dtype=np.dtype(dtype), mask=mask, name=name
         )
     elif np.dtype(dtype).type in (np.object_, np.str_):
         if not isinstance(buffer, nvstrings.nvstrings):
             raise TypeError
-        return string.StringColumn(data=buffer)
+        return string.StringColumn(data=buffer, name=name)
     else:
-        return numerical.NumericalColumn(data=buffer, dtype=dtype, mask=mask)
+        return numerical.NumericalColumn(
+            data=buffer, dtype=dtype, mask=mask, name=name
+        )
 
 
-def as_column(arbitrary, nan_as_null=True, dtype=None):
+def as_column(arbitrary, nan_as_null=True, dtype=None, name=None):
     """Create a Column from an arbitrary object
 
     Currently support inputs are:
@@ -259,6 +262,9 @@ def as_column(arbitrary, nan_as_null=True, dtype=None):
     from cudf.dataframe.index import Index
     from cudf.bindings.cudf_cpp import np_to_pa_dtype
 
+    if name is None and hasattr(arbitrary, "name"):
+        name = arbitrary.name
+
     if isinstance(arbitrary, Column):
         categories = None
         if hasattr(arbitrary, "categories"):
@@ -272,7 +278,6 @@ def as_column(arbitrary, nan_as_null=True, dtype=None):
 
     elif isinstance(arbitrary, Series):
         data = arbitrary._column
-
     elif isinstance(arbitrary, Index):
         data = arbitrary._values
 
@@ -502,7 +507,8 @@ def as_column(arbitrary, nan_as_null=True, dtype=None):
                         np.array(arbitrary, dtype=np_type),
                         nan_as_null=nan_as_null,
                     )
-
+    if hasattr(data, "name") and (name is not None):
+        data.name = name
     return data