+++ python/cudf/cudf/dataframe/string.py
@@ -58,7 +58,9 @@ class StringMethods(object):
                     ret = getattr(self._parent._data, attr)(*args, **kwargs)
                     if isinstance(ret, nvstrings.nvstrings):
                         ret = Series(
-                            columnops.as_column(ret), index=self._index
+                            columnops.as_column(ret),
+                            index=self._index,
+                            name=self._parent.name,
                         )
                     return ret
 
@@ -90,7 +92,7 @@ class StringMethods(object):
         column = columnops.build_column(
             Buffer(out_dev_arr), np.dtype("int32"), mask=mask
         )
-        return Series(column, index=self._index)
+        return Series(column, index=self._index, name=self._parent.name)
 
     def cat(self, others=None, sep=None, na_rep=None):
         """
@@ -202,6 +204,7 @@ class StringMethods(object):
         out = Series(
             self._parent.data.cat(others=others, sep=sep, na_rep=na_rep),
             index=self._index,
+            name=self._parent.name,
         )
         if len(out) == 1 and others is None:
             out = out[0]
@@ -251,7 +254,7 @@ class StringMethods(object):
 
         out = self._parent.data.extract(pat)
         if len(out) == 1 and expand is False:
-            return Series(out[0], index=self._index)
+            return Series(out[0], index=self._index, name=self._parent.name)
         else:
             out_df = DataFrame(index=self._index)
             for idx, val in enumerate(out):
@@ -308,7 +311,7 @@ class StringMethods(object):
             Buffer(out_dev_arr), np.dtype("bool"), mask=mask
         )
 
-        return Series(column, index=self._index)
+        return Series(column, index=self._index, name=self._parent.name)
 
     def replace(self, pat, repl, n=-1, case=None, flags=0, regex=True):
         """
@@ -353,6 +356,7 @@ class StringMethods(object):
         return Series(
             self._parent.data.replace(pat, repl, n=n, regex=regex),
             index=self._index,
+            name=self._parent.name,
         )
 
     def lower(self):
@@ -366,7 +370,11 @@ class StringMethods(object):
         """
         from cudf.dataframe import Series
 
-        return Series(self._parent.data.lower(), index=self._index)
+        return Series(
+            self._parent.data.lower(),
+            index=self._index,
+            name=self._parent.name,
+        )
 
     def split(self, pat=None, n=-1, expand=True):
         """
@@ -413,7 +421,7 @@ class StringColumn(columnops.TypedColumnBase):
     """Implements operations for Columns of String type
     """
 
-    def __init__(self, data, null_count=None, **kwargs):
+    def __init__(self, data, null_count=None, name=None, **kwargs):
         """
         Parameters
         ----------
@@ -429,6 +437,7 @@ class StringColumn(columnops.TypedColumnBase):
         assert isinstance(data, nvstrings.nvstrings)
         self._data = data
         self._dtype = np.dtype("object")
+        self._name = name
 
         if null_count is None:
             null_count = data.null_count()
@@ -571,7 +580,7 @@ class StringColumn(columnops.TypedColumnBase):
 
     def to_pandas(self, index=None):
         pd_series = self.to_arrow().to_pandas()
-        return pd.Series(pd_series, index=index)
+        return pd.Series(pd_series, index=index, name=self.name)
 
     def to_array(self, fillna=None):
         """Get a dense numpy array for the data.