+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Improvements
 
+- PR #2109 added name to Python column classes
 - PR #1947 Cleanup serialization code
 - PR #2125 More aggregate in java API
 - PR #2131 Chunk rows logic added to csv_writer